Fix: nav 유저 정보 생성, 페이지 정리
