+++ next-sprint/pages/_app.js
@@ -4,6 +4,7 @@ import Nav from '@/components/Nav';
 import Footer from '@/components/Footer';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import React from 'react';
+import { UserProvider } from '@/contexts/UserContext';
 
 export default function App({ Component, pageProps }) {
   const [queryClient] = React.useState(
@@ -27,7 +28,7 @@ export default function App({ Component, pageProps }) {
     ));
   return (
     <QueryClientProvider client={queryClient}>
-      {getLayout(<Component {...pageProps} />)}
+      <UserProvider>{getLayout(<Component {...pageProps} />)}</UserProvider>
     </QueryClientProvider>
   );
 }
